bump aiida-pythonjob to 0.1.8
