+++ pyproject.toml
@@ -22,7 +22,7 @@ classifiers = [
 dependencies = [
   'aiida-core>=2.3',
   'aiida-quantumespresso>=4.8',
-  'aiida-pythonjob'
+  'aiida-pythonjob>=0.1.8'
 ]
 dynamic = ['description', 'version']
 keywords = ['aiida', 'workflows', 'core-level', 'spectroscopy']
